+++ test/support/modules/bicycle_subclass_test.rb
@@ -1,0 +1,13 @@
+module BicycleSubclassTest
+  def test_responds_to_post_initialize
+    assert_respond_to(@object, :post_initialize)
+  end
+
+  def test_responds_to_local_spares
+    assert_respond_to(@object, :local_spares)
+  end
+
+  def test_responds_to_default_tire_size
+    assert_respond_to(@object, :default_tire_size)
+  end
+end